+++ example/outputs.tf
@@ -1,18 +1,34 @@
 ######################
 # S3 Bucket Outputs: #
 ######################
-output "bucket_id" {
-  value = module.demo_s3bucket.id
+output "id" {
+  value = module.defaults.id
 }
 
-output "bucket_arn" {
-  value = module.demo_s3bucket.arn
+output "arn" {
+  value = module.defaults.arn
 }
 
 output "bucket_domain_name" {
-  value = module.demo_s3bucket.domain_name
+  value = module.defaults.domain_name
 }
 
-output "bucket_region" {
-  value = module.demo_s3bucket.region
+output "bucket_regional_domain_name" {
+  value = module.defaults.bucket_regional_domain_name
+}
+
+output "region" {
+  value = module.defaults.region
+}
+
+output "website_endpoint" {
+  value = module.defaults.website_endpoint
+}
+
+output "website_domain" {
+  value = module.defaults.website_domain
+}
+
+output "hosted_zone_id" {
+  value = module.defaults.hosted_zone_id
 }